Use uint64_t with gazebo-entity user dataa (#1451)

+++ src/gui/plugins/transform_control/TransformControl.cc
@@ -801,8 +801,7 @@ void TransformControlPrivate::HandleTransform()
         auto entityId = kNullEntity;
         try
         {
-          entityId = static_cast<unsigned int>(
-            std::get<int>(visual->UserData("gazebo-entity")));
+          entityId = std::get<uint64_t>(visual->UserData("gazebo-entity"));
         }
         catch (std::bad_variant_access &)
         {
